requirements.txt.txt
Django==4.2.7
tensorflow==2.15.0
Pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
opencv-python==4.8.1.78
gunicorn==21.2.0
whitenoise==6.6.0